requirements.txt
Django==3.2.5
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-yasg==1.20.0
Markdown==3.3.4
Pillow==8.3.1